+++ src/components/landing-page/Events.jsx
@@ -32,7 +32,7 @@ return (
     <div>
     
 
-    <div class="album py-5 bg-light">
+    <div class="album py-5">
     <div class="container">
 
       <div class="row row-cols-1 row-cols-sm-2 row-cols-md-4 g-4">
