+++ src/components/landing-page/NavBar.jsx
