+++ src/components/landing-page/mockup-events.json
