+++ src/components/landing-page/styles/categories.css
@@ -12,14 +12,16 @@
     font-weight: 600;
     color: #39364F;
     font-family: Montserrat, sans-serif;
-    margin-left: 110px;
+    /* margin-left: 110px; */
     /* margin-top: 20px; */
     margin-bottom: 20px;
 }
 
 .location{
     display: flex;
+    margin-top: 40px;
 }
+
 
 
 .location--text{
@@ -27,7 +29,7 @@
     color: #39364F;
     font-family: Montserrat, sans-serif;
     margin-left: 10px;
-
+    
 }
 
 .categories-tabs{
@@ -35,4 +37,27 @@
     justify-content: center;
     align-items: center;
     margin-bottom: 20px;
+    margin-top: 10px;
 }
+
+
+.nav-item a {
+    color: #75788C;
+    font-weight: 600;
+  }
+
+  .trending-categories h4{
+    margin-top: 30px;
+    margin-bottom: 30px;
+    font-weight: 600;
+    color: #39364F;
+    font-family: Montserrat, sans-serif;
+}
+  .events-in h4{
+    margin-top: 30px;
+    font-weight: 600;
+    color: #39364F;
+    font-family: Montserrat, sans-serif;
+}
+
+
