+++ src/components/landing-page/styles/events.css
@@ -15,5 +15,5 @@
 
 
 .event-card:hover{
-    box-shadow: 0px 20px 40px rgba(0, 0, 0, 0.2);
+    box-shadow: 0px 20px 40px rgba(0, 0, 0, 0.3);
 }
